fragments of mediateka screen were moved to own package
