+++ app/src/main/java/bes/max/playlistmaker/presentation/mediateka/MediatekaViewPagerAdapter.kt
@@ -4,6 +4,8 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.lifecycle.Lifecycle
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import bes.max.playlistmaker.presentation.mediateka.fragments.FavoriteTracksFragment
+import bes.max.playlistmaker.presentation.mediateka.fragments.PlaylistFragment
 
 class MediatekaViewPagerAdapter(fragmentManager: FragmentManager, lifecycle: Lifecycle) : FragmentStateAdapter(fragmentManager, lifecycle) {
     override fun getItemCount(): Int = 2
