+++ app/src/main/java/bes/max/playlistmaker/presentation/mediateka/fragments/PlaylistFragment.kt
@@ -1,25 +1,26 @@
-package bes.max.playlistmaker.presentation.mediateka
+package bes.max.playlistmaker.presentation.mediateka.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import bes.max.playlistmaker.R
-import bes.max.playlistmaker.databinding.FragmentFavoriteTracksBinding
-import bes.max.playlistmaker.presentation.mediateka.viewmodels.FavoriteTracksViewModel
+import bes.max.playlistmaker.databinding.FragmentPlaylistBinding
+import bes.max.playlistmaker.presentation.mediateka.viewmodels.PlaylistViewModel
 import bes.max.playlistmaker.presentation.utils.BindingFragment
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class FavoriteTracksFragment : BindingFragment<FragmentFavoriteTracksBinding>() {
+class PlaylistFragment : BindingFragment<FragmentPlaylistBinding>() {
 
-    private val favoriteTracksViewModel: FavoriteTracksViewModel by viewModel()
+    private val playlistViewModel: PlaylistViewModel by viewModel()
 
     override fun createBinding(
         inflater: LayoutInflater,
         container: ViewGroup?
-    ): FragmentFavoriteTracksBinding {
-        return FragmentFavoriteTracksBinding.inflate(inflater, container, false)
+    ): FragmentPlaylistBinding {
+        return FragmentPlaylistBinding.inflate(inflater, container, false)
     }
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -27,17 +28,16 @@ class FavoriteTracksFragment : BindingFragment<FragmentFavoriteTracksBinding>() 
     }
 
 
-
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_favorite_tracks, container, false)
+        return inflater.inflate(R.layout.fragment_playlist, container, false)
     }
 
     companion object {
 
-        fun newInstance() = FavoriteTracksFragment()
+        fun newInstance() = PlaylistFragment()
 
     }
