+++ app/src/main/res/layout/activity_mediateka.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/mediateka_fragment_container"
-    android:name="bes.max.playlistmaker.presentation.mediateka.MediatekaFragment"
+    android:name="bes.max.playlistmaker.presentation.mediateka.fragments.MediatekaFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent" />
