+++ app/src/main/res/layout/fragment_favorite_tracks.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.mediateka.FavoriteTracksFragment">
+    tools:context=".presentation.mediateka.fragments.FavoriteTracksFragment">
 
     <androidx.constraintlayout.widget.Group
         android:id="@+id/fragment_favorite_tracks_placeholder"
