+++ app/src/main/res/layout/fragment_playlist.xml
@@ -5,7 +5,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.mediateka.PlaylistFragment">
+    tools:context=".presentation.mediateka.fragments.PlaylistFragment">
 
     <androidx.constraintlayout.widget.Group
         android:id="@+id/fragment_playlist_placeholder"
